% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps.R
\name{ps}
\alias{ps}
\title{Print script information}
\usage{
ps(file = dir()[grep("\\\\.R$", dir(), ignore.case = TRUE)][1])
}
\arguments{
\item{method2order}{}

\item{cex}{}

\item{mar}{}
}
\value{
None
}
\description{
Print script information
}
\examples{
ps()

}
