% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s.R
\name{s}
\alias{s}
\title{source the n-th section of the script, defaulting to the last section ####}
\usage{
s(
  start = "last",
  end = start,
  file = dir()[grep("\\\\.R$", dir(), ignore.case = TRUE)][1]
)
}
\arguments{
\item{method2order}{}

\item{cex}{}

\item{mar}{}
}
\value{
None
}
\description{
source the n-th section of the script, defaulting to the last section ####
}
\examples{
s()

}
