% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/so.R
\name{so}
\alias{so}
\title{source only a single line or section, defaulting to the first section ####}
\usage{
so(beg = "", file = dir()[grep("\\\\.R$", dir(), ignore.case = TRUE)][1])
}
\arguments{
\item{method2order}{}

\item{cex}{}

\item{mar}{}
}
\value{
None
}
\description{
source only a single line or section, defaulting to the first section ####
}
\examples{
so()

}
