% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss.R
\name{ss}
\alias{ss}
\title{source the n-th section of the script, defaulting to the last section ####}
\usage{
ss(
  beg = "last",
  end = beg,
  file = dir()[grep("\\\\.R$", dir(), ignore.case = TRUE)][1]
)
}
\arguments{
\item{method2order}{}

\item{cex}{}

\item{mar}{}
}
\value{
None
}
\description{
source the n-th section of the script, defaulting to the last section ####
}
\examples{
ss()

}
