% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s.R
\name{s}
\alias{s}
\title{Any comment line which includes at least four trailing dashes (-), 
equal signs (=), or pound signs (#) automatically creates a code section.}
\usage{
s(
  beg = 1,
  end = as.numeric(length(readLines(file, warn = FALSE))),
  file = dir()[grep("\\\\.R$", dir(), ignore.case = TRUE)][1],
  number.is.section = FALSE
)
}
\arguments{
\item{method2order}{}

\item{cex}{}

\item{mar}{}
}
\value{
None
}
\description{
Any comment line which includes at least four trailing dashes (-), 
equal signs (=), or pound signs (#) automatically creates a code section.
}
\examples{
s(1, 10)

}
